third_party/blink/web_tests/fast/files/file-reader-sandbox-iframe.html

<html>
<head>
<script>
if (window.testRunner) {
    testRunner.setAllowFileAccessFromFileURLs(false);
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
</script>
</head>
<body>
<iframe src="resources/file-reader-real.html" sandbox="allow-scripts"></iframe>
<div id="results"></div>
Test that FileReader works under the sandboxed iframe.
</body>
</html>
