third_party/blink/web_tests/fast/files/file-reader-sandbox-iframe-expected.txt

CONSOLE MESSAGE: Received data: 0123456789
Test that FileReader works under the sandboxed iframe.
